fixed some computation errors

+++ src/main/scala/Euler5.scala
@@ -1,9 +1,7 @@
 package org.draegisoft.euler
 
-import collection.mutable.Seq
-
 object Euler5 {
-  def euler(lower: Int, upper: Int): Long = {
-    (1 /: (lower to upper)) (lcm(_, _))
+  def euler(lower: Long, upper: Long): Long = {
+    (1L /: (lower to upper)) (lcm(_, _))
   }
 }
